experiment/requirements.txt
# MAIF High-Performance Implementation Requirements
# Core dependencies for production deployment

# Async I/O and performance
uvloop>=0.17.0

# Compression and serialization
lz4>=4.3.2
msgpack>=1.0.5
zstandard>=0.21.0

# Cryptography and security
cryptography>=41.0.0
PyJWT>=2.8.0
bcrypt>=4.0.1
pynacl>=1.5.0

# Data processing and ML
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# HTTP and networking
aiohttp>=3.8.0
httpx>=0.24.0

# Monitoring and observability
psutil>=5.9.0
structlog>=23.1.0

# Configuration and utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
click>=8.1.0

# Production storage backends (uncomment as needed for deployment)
# Redis cluster for hot cache
# redis>=4.5.0
# aioredis>=2.0.0

# Kafka for write-ahead logging
# aiokafka>=0.8.0
# kafka-python>=2.0.0

# FoundationDB for metadata
# foundationdb>=7.1.0

# Vector database for embeddings
# qdrant-client>=1.6.0
# weaviate-client>=3.24.0

# Object storage
# boto3>=1.26.0
# minio>=7.1.0

# Analytics database
# duckdb>=0.9.0
# clickhouse-driver>=0.2.0

# Graph database for knowledge graphs
# neo4j>=5.12.0
# rdflib>=7.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0
black>=23.0.0
mypy>=1.5.0
flake8>=6.0.0
isort>=5.12.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0 